a bit of reorg
