+++ pl_bolts/datasets/cifar10_dataset.py
@@ -7,6 +7,7 @@ import torch
 from torch import Tensor
 
 from pl_bolts.datasets import LightDataset
+from pl_bolts.datasets.utils import safe_extract_tarfile
 from pl_bolts.utils import _PIL_AVAILABLE
 from pl_bolts.utils.stability import under_review
 from pl_bolts.utils.warnings import warn_missing_pkg
@@ -118,26 +119,7 @@ class CIFAR10(LightDataset):
     def _extract_archive_save_torch(self, download_path):
         # extract achieve
         with tarfile.open(os.path.join(download_path, self.FILE_NAME), "r:gz") as tar:
-
-            def is_within_directory(directory, target):
-
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-
-                return prefix == abs_directory
-
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-
-                tar.extractall(path, members, numeric_owner=numeric_owner)
-
-            safe_extract(tar, path=download_path)
+            safe_extract_tarfile(tar, path=download_path)
         # this is internal path in the archive
         path_content = os.path.join(download_path, "cifar-10-batches-py")
