+++ pl_bolts/datasets/imagenet_dataset.py
@@ -1,21 +1,16 @@
-import gzip
 import hashlib
 import os
 import shutil
-import sys
-import tarfile
 import tempfile
-import zipfile
 from contextlib import contextmanager
 
 import numpy as np
 import torch
 
+from pl_bolts.datasets.utils import extract_archive
 from pl_bolts.utils import _TORCHVISION_AVAILABLE
 from pl_bolts.utils.stability import under_review
 from pl_bolts.utils.warnings import warn_missing_pkg
-
-PY3 = sys.version_info[0] == 3
 
 if _TORCHVISION_AVAILABLE:
     from torchvision.datasets import ImageNet
@@ -247,116 +242,3 @@ def parse_devkit_archive(root, file=None):
         META_FILE = "meta.bin"
 
         torch.save((wnid_to_classes, val_wnids), os.path.join(root, META_FILE))
-
-
-@under_review()
-def extract_archive(from_path, to_path=None, remove_finished=False):
-    if to_path is None:
-        to_path = os.path.dirname(from_path)
-
-    PY3 = sys.version_info[0] == 3
-
-    if _is_tar(from_path):
-        with tarfile.open(from_path, "r") as tar:
-
-            def is_within_directory(directory, target):
-
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-
-                return prefix == abs_directory
-
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-
-                tar.extractall(path, members, numeric_owner=numeric_owner)
-
-            safe_extract(tar, path=to_path)
-    elif _is_targz(from_path):
-        with tarfile.open(from_path, "r:gz") as tar:
-
-            def is_within_directory(directory, target):
-
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-
-                return prefix == abs_directory
-
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-
-                tar.extractall(path, members, numeric_owner=numeric_owner)
-
-            safe_extract(tar, path=to_path)
-    elif _is_tarxz(from_path) and PY3:
-        # .tar.xz archive only supported in Python 3.x
-        with tarfile.open(from_path, "r:xz") as tar:
-
-            def is_within_directory(directory, target):
-
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-
-                return prefix == abs_directory
-
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-
-                tar.extractall(path, members, numeric_owner=numeric_owner)
-
-            safe_extract(tar, path=to_path)
-    elif _is_gzip(from_path):
-        to_path = os.path.join(to_path, os.path.splitext(os.path.basename(from_path))[0])
-        with open(to_path, "wb") as out_f, gzip.GzipFile(from_path) as zip_f:
-            out_f.write(zip_f.read())
-    elif _is_zip(from_path):
-        with zipfile.ZipFile(from_path, "r") as z:
-            z.extractall(to_path)
-    else:
-        raise ValueError(f"Extraction of {from_path} not supported")
-
-    if remove_finished:
-        os.remove(from_path)
-
-
-@under_review()
-def _is_targz(filename):
-    return filename.endswith(".tar.gz")
-
-
-@under_review()
-def _is_tarxz(filename):
-    return filename.endswith(".tar.xz")
-
-
-@under_review()
-def _is_gzip(filename):
-    return filename.endswith(".gz") and not filename.endswith(".tar.gz")
-
-
-@under_review()
-def _is_tar(filename):
-    return filename.endswith(".tar")
-
-
-@under_review()
-def _is_zip(filename):
-    return filename.endswith(".zip")
